+++ src/source.extension.cs
@@ -9,7 +9,7 @@ namespace LessCompiler
     {
         public const string Id = "d32c5250-fa82-4da6-9732-5518fabebfef";
         public const string Name = "LESS Compiler";
-        public const string Description = @"An alternative LESS compiler with little to no setup. Uses the node.js based LESS compiler under the hood with AutoPrefixer and CSSComb built in.";
+        public const string Description = @"An alternative LESS compiler with no setup. Uses the official node.js based LESS compiler under the hood with AutoPrefixer and CSSComb built in.";
         public const string Language = "en-US";
         public const string Version = "0.6";
         public const string Author = "Mads Kristensen";